+++ tests/test_geometamaker.py
@@ -59,11 +59,11 @@ def create_vector(target_filepath, field_map=None, driver='GEOJSON'):
 
 def create_raster(
         numpy_dtype, target_path,
-        pixel_size=(1, 1), projection_epsg=4326,
+        pixel_size=(1, 1), raster_size=(2, 2), projection_epsg=4326,
         origin=(0, 0), n_bands=2, define_nodata=True):
     driver_name, creation_options = DEFAULT_GTIFF_CREATION_TUPLE_OPTIONS
     raster_driver = gdal.GetDriverByName(driver_name)
-    ny, nx = (2, 2)
+    nx, ny = raster_size
     gdal_type = gdal_array.NumericTypeCodeToGDALTypeCode(numpy_dtype)
     raster = raster_driver.Create(
         target_path, nx, ny, n_bands, gdal_type)
@@ -78,7 +78,7 @@ def create_raster(
     if projection_wkt is not None:
         raster.SetProjection(projection_wkt)
 
-    base_array = numpy.full((2, 2), 1, dtype=numpy_dtype)
+    base_array = numpy.full((nx, ny), 1, dtype=numpy_dtype)
 
     target_nodata = pygeoprocessing.choose_nodata(numpy_dtype)
 
@@ -315,11 +315,34 @@ class GeometamakerTests(unittest.TestCase):
              'STATISTICS_VALID_PERCENT': '100'})
 
     def test_describe_raster_band_compute_statistics(self):
-        """Test band statistics will be included if they already exist."""
+        """Test band statistics will be computed if they do not already exist."""
         import geometamaker
 
         datasource_path = os.path.join(self.workspace_dir, 'raster.tif')
         create_raster(numpy.int16, datasource_path, n_bands=1)
+
+        resource = geometamaker.describe(datasource_path, compute_stats=True)
+        self.assertEqual(
+            resource.data_model.bands[0].gdal_metadata,
+            {'STATISTICS_MINIMUM': '1',
+             'STATISTICS_MAXIMUM': '1',
+             'STATISTICS_MEAN': '1',
+             'STATISTICS_STDDEV': '0',
+             'STATISTICS_VALID_PERCENT': '100'})
+
+    def test_describe_raster_band_compute_statistics_valid_percent(self):
+        """Test band statistics will be computed if VALID_PERCENT is missing."""
+        import geometamaker
+
+        datasource_path = os.path.join(self.workspace_dir, 'raster.tif')
+        create_raster(numpy.int16, datasource_path, n_bands=1)
+        raster = gdal.OpenEx(datasource_path)
+        band = raster.GetRasterBand(1)
+        # Deliberately set stats metadata that do not include VALID_PERCENT
+        band.SetMetadataItem('STATISTICS_MINIMUM', '1')
+        band.SetMetadataItem('STATISTICS_MAXIMUM', '1')
+        band.SetMetadataItem('STATISTICS_MEAN', '1')
+        band.SetMetadataItem('STATISTICS_STDDEV', '0')
 
         resource = geometamaker.describe(datasource_path, compute_stats=True)
         self.assertEqual(